requirements.txt
fastapi
uvicorn 
sqlalchemy 
alembic 
psycopg2-binary
python-dotenv
pydantic[email]
email-validator